LCM: Add parsing for ACE parameters.
The current structure of the ACE material parameters is
difficult to parse with the existing parsing model in the
constitutive model parameters class.

Will attempt another way of structuring the ACE parameters so that
reading them from the YAML file is easier. Check this code anyway
in case we change our minds later and decide to reuse it.

Pair programming with Jennifer Frederick.