Added the option of scattering a scalar nodal parameter only once.
This is useful when loading (or computing) a field that is constant during
the simulation, and that we want to use as a DirichletField. Currently,
DirichletField grabs the datum from the DistParamLib. So we need to scatter
the field we loaded (or computed), but since it is constant over the whole
simulation, it does not make sense to scatter it every time. Once is enough.

Notice that, if a problem is allowing this case, the constructEvaluators
method should add the scatter tag as a requested field. This will make sure
that, during the first residual evaluation, the field is scattered, so that
is then available when the Dirichlet field manager is evaluated (recall
that the volume field manager is evaluated BEFORE the dirichlet one).