Added the option of computing field from math expr in GenericSTKMeshStruct
When loading the field, if 'Field Expression' is present, it is used to
compute the field, based on the xyz coordinates of the nodes. Only nodal
fields are allowed, and no layered fields. The user can supply an arbitrary
list of parameters in the input files, which will be used to evaluate the
field. E.g.,

<Parameter name="Field Expression" type="Array(string)" value="{a=2.0,a*x*x}"/>