[#292] Removed memoizer macro and added input parameter
The MDField memoizer is now disabled by default and must be enabled during
the construction of the evaluator. This allows for flexibility when
determining whether memoization is required in a specific evaluator for
a given problem.

The memoizer can be controlled through the use of an input parameter
in the "Problem" parameter list. The parameter is currently only
implemented in FELIX FOStokes:
<Parameter name="Use MDField Memoization" type="bool" value="true"/>