FELIX added/fixed penalization for solution bounds in hydrology problem
For mass equation, we already had penalization (concentrated at nodes) for
overshoots (phi>phi_0=rhoi*g*H). Now we also have it for undershoots (phi<0),
and we can also set a penalization coefficient (same for both). Also, the
penalization is an integral term, so, e.g., |min(0,phi)|^2 is integrated
over the element.

For cavities equation, we added penalization for negative thickness. Again,
the user can specify a penalization coefficient.