LCM: Fix initialization of solutions for Schwarz.
Albany::ModelEvaluator::getNominalValues() returns a copy
of its Thyra::ModelEvaluatorBase::InArgs nominal values object.

This object is used to hold initial values for various vectors,
including the solution vector. But because the accessor returns a copy
instead of a reference, setting non-default initial values has no
effect.

It could be modified to return a pointer or reference so that the
non-default initial values persist. But I suspect that would break a
lot of code.

Instead, create the corresponding mutator setNominalValues. Thus, a
Thyra::ModelEvaluatorBase::InArgsSetup object is created and desired
initial values are assigned to it. Note that this object is of type
InArgsSetup instead of InArgs. Not sure why the difference, but
using an InArgs object to initialize does not work either.

Once this InArgsSetup is created and desired initial values are
assigned to it, then we use the newly created mutator setNominalValues
to successfully set arbitrary initial values for the solver.

This seems to fix some of the puzzling behavior in Schwarz where
convergence of the algorithm became harder and harder to achieve the
farther it went from the initial, default solution.