Extending "Scale BC Dofs: true" scaling option to work for non-constant scaling.
In this case, we get either the diagonal value of the Jacobian,
or the absolute value of the rowsum, compute the maximum, and
scale the DBCs by this value.  As the Jacobian evolves, so does
the scaling; therfore it is not necessary to specify a priori
a scaling value in the input file.

A bug was discovered in the DBC scaling today (see issue #307)
that should be fixed prior to expecting the new option
to work robustly for all problems.