Bug fixes to PHAL_DOFCellToSideQP evaluator
The layouts of PHX fields are not set in the constructor, but in during
the setFieldData calls. Also, the dimensions of the cell field are useless,
and the number of nodes was erroneously grabbed as the 2nd dimension of the
side field (probably cut-and-paste from the DOFCellToSide evaluator).
Also, the scalar type of the basis functions is RealType, not EvalT::ScalarT.