Declaring Albany comm macros only once.
Albany_Utils.hpp and Petra_Converters.hpp were both defining the macro
Albany_MPI_Comm and others. Even though they both defined it to the same
thing, this is not 100% standard (some compiler may issue a warning),
and potentially dangerous if one of the two changes.

Hence, moved the define's to Albany_DataTypes.hpp, which is included by both.