Added option to enable CMP0060 policy new behavior
This policy's new behavior makes cmake honor the full path of libraries when given,
instead of using the common '-L/path/to -llibname' split. This helps when the user has some
libraries in system folder that have the same name as libraries needed (directly
or indirectly) by Albany. E.g., the user may have a system installation of netcdf,
like /usr/lib/libnetcdf.so, which would shadow an ad-hoc installation in another
folder, even if its location was specifically passed to cmake. With the new behavior,
if cmake is given a full path to a library, say, /my/path/libnetcdf.so, it won't split
it into '-L/my/path -lnetcdf', but link it directly against
'/my/path/libnetcdf.so'.