Removed calls to deprecated function in Trilinos
The function KokkosClassic::Details::getNode<Node>() is deprecated now.
In Albany, it was used in the Petra converters, to create a node for the
Tpetra objects. However, this is pointless, cause the constructor of
Tpetra::Map allows to not pass a node. That makes sense also because
the node is never used, except for its types, so there is no need to
pass one (the node type is already part of the template parameters of
Tpetra::Map).

Hence, I removed the calls to KokkosClassic::Details::getNode.
Additionally, I removed the last optional argument in the Petra converters,
which was a pointer to a node object. Nowhere in Albany that argument
was passed a value, and we always relied on the default (which, again,
was not really used in the Tpetra::Map constructor).