Replacing tpetra/epetra const (multi)vectors in Workset with Thyra ones
This is a first step of the Thyra refactor process. The goal is to get
rid of all *petra stuff in the high level classes, delegating the
conversion to the proper type to the low level ones (gather/scatter
and read/write).

As a first step, I replaced the solution vectors x, xdot, and xdotdot,
as well as the tangent multivectors Vx, Vxdot, Vxdotdot, and Vp in
the workset with Thyra (multi)vectors. This means that every object
that tries to access these members of Workset, need to access them
as Thyra objects, and, if needed, they will convert them to the proper
*petra type. This is currently generating quite a bit of conversion
calls. However, the plan is to move the conversions up the stack call,
until we reach the ModelEvaluator level, where currently the Thyra
objects are converted to *petra objects. Once we reach this point,
we will eliminate the need of conversion on the upper end of the stack
call. The low level classes (gather/scatter and read/write) will be left
as the only ones(*) that need to explicitly cast to a concrete linear
algebra type.

(*) There is often a need ot combine/scatter vectors from the unique
    to the overlapped distribution. This is also a point of the code
    where it is necessary to know what the underlying linear algebra
    object looks like. For now, I moved this to the solution manager
    level, but it may change.