Helper functions to get Thyra objects from Tpetra objects (and viceversa)
The Thyra conversion routines require that the input RCP be nonnull.
This requires us to keep checking if a ptr is null before calling the
conversion routine. Instead, these convenience routines do this for
us, so we do not pollute our code with if statements; if the input
pointer is null, they return a null pointer.

Also, added typedef for Thyra Tpetra vector space