LCM: Set initial guess for Newmark solve in Dynamic Schwarz loop
Using the newly implemented tool to set the initial guess for the
newmark loop on each schwarz iteration. It uses the displacements from
the previous converged schwarz iteration. Newmark used to
automatically use the displacements from the previous converged
loadstep as an initial guess, but the previous schwarz iterations
provide an initial guess that should be closer to the converged
solution.