LCM: Cherry-picking relevant commits from tempus-lcm-schwarz branch into master, resolving conflicts.
******************************

LCM: adding routine to set initial guess for Schwarz

the routine is meant to be called before entering the schwarz loop. It
computes an explicit update of the displacements, and applies those
displacements to each Schwarz domain. This was shown to be a good
initial guess for small (near the stable explicit timestep) timesteps
in simple 1D cases, but it is not so good for larger timesteps, which
is why it is not being used.