LCM ACE: Use ACE temperature for models.
The ACE models use a separate temperature field termed
ACE temperature. The reason is that the evaluator for the
ACE temperature residual requires special features to account
for the phase change of water and other related mechanisms.

To avoid polluting the standard temperature residual, it seemed
better to use a separate temperature field exclusive to the ACE
models that require the ACE temperature residual.

Switch the ACE models to use this ACE temperature, and update
the ACE permafrost model to incorporate changes that existed
in the ACE ice model only.