LCM: Clang-format almost everything. [#38]
Use clang-format to format almost all the code under the LCM
subdirectory. This uncovered some inconsistencies with #include
statements that were fixed.

The main exception to this reformatting is the crystal plasticity
model. Jay Foulk, Coleman Alleman, Guy Bergel, and Dave Littlewood
would gang up and kill me if I touch that code in this manner.