Bug fixes in related to thyra vector's initialization.
Thyra does NOT guarantee that the method createMember (or createMembers)
of Thyra::VectorSpaceBase will initialize the (multi)vector to 0. In fact,
the TpetraVectorSpace class explicitly asks the Tpetra constructors to NOT
zero-out the vector(s). Even more: if the macro THYRA_INITIALIZE_VECS_MULTIVECS_WITH_NANS
is defined, Thyra will make sure the vectors are init-ed with nans (which
is indeed better than random numbers).

In some places this is ok (e.g., responses are explicitly init-ed to 0
before being filled), in other places this was dubious or wrong, so we
added explicit calls to v->assign(0.0).