Bug fix in Petra converters
When creating an array view from Epetra's vector, to be fed to Tpetra::Vector's
constructor, do not say that the length of the view is the global number of
elements.

This is innocuous if the rank stores at least one element: in this case,
the pointer extracted from Epetra's constructor is nonnull, and the array
view is legally constructed.
However, in the (very peculiar) case of no entries stored in the current
rank, Epetra stores a null pointer for the values; the array view constructor
then sees a null pointer but a non-zero size for the view (since we were
passing the global length), and complains.

By passing the local length (as it is already done for MV, btw), in the case
with 0 entries on rank, the constructor for the array view sees a null pointer
and a zero lenght, which is legit.