LCM: Improve convergence criterion for dynamic Schwarz.
The convergence criterion for dynamic Schwarz can be
summarized as:

y := d + dt * v + dt^2 * a
dy := dd + dt * dv + dt^2 * da
ae := dy
re := dy / y

where:

d : displacement
v : velocity
a : acceleration
dd: displacement difference
dv: velocity difference
da: acceleration difference
dt: current time step
ae: absolute error
re: relative error

The Schwarz iteration is considered converged if the relative
error OR the absolute error are less or equal that corresponding
tolerances.

This criterion has at least two problems: it changes with the time step,
and it does not provide a mechanism for giving more or less
weight to the velocity or acceleration in the convergence.

Instad of providing separate tolerances for the displacement,
velocity, and acceleration, the criterion is changed to:

y := d + fv * v + fa * a
dy := dd + fv * dv + fa * da

where

fv: tolerance factor for velocity
fa: tolerance factor for acceleration

and the rest as before. By choosing fv and fa judiciously, one can
achieve the same effect as providing separate tolerances for the
velocity and acceleration. The factors fv and fa are initialized to
the *initial* time step and its square, respectively, but can
otherwise be changed to arbitrary values in the input. This provides
a way to give more weight to the velocity and acceleration if desired.