LCM: Make Schwarz convergence stricter.
We previously changed the convergence criterion for dynamic Schwarz
such that the velocity and acceleration have scaling factors. This
was introduced so that convergence requires tight matches in
velocity or acceleration.

The convergence criterion is essentially:

abs_error <= abs_tol || rel_error <= rel_tol

Drop test calculations show that this is not enough to prevent
divergence from spurious waves that are amplified. The reason
is that accelerations can grow very large and they will be
considered converged as long as they satisfy the relative error
criterion. To avoid this, the criterion is changed to

abs_error <= abs_tol && rel_error <= rel_tol

so that now, any Schwarz run MUST satisfy both to be considered
converged.