Bug fix in the Tpetra CAS manager
It can happen that one creates a Thyra_Vector from a Tpetra_Vector, then
feeds the thyra vector to a routine expecting a Thyra_MultiVector. When
we then try to get a Tpetra_MultiVector from the thyra mv, we get an error,
since Thyra::TpetraVector does not inherit from Thyra::TpetraMultiVector.

So when we receive a Thyra_MV, we have to try to get a tpetra object twice:
the first time we try with a MV, then, if we fail, we try to get a vector.