LCM ACE: Yield stress function of ice saturation.
Add a preliminary linear dependence of both the ice and permafrost
yield stress on the ice saturation, which in turns depends on
temperature and salinity.

For the ice, the effective yield stress is simply the product
of the ice saturation and the crystalline phase yield stress.

For the permafrost, a minimum yield stress needs to be specified
separately, and the effective yield stress varies linearly
between this value and the value of the fully frozen material.

As more data comes in from mechanical charaterization experiments,
this linear dependence assumption may prove to be too simplistic,
but it can be changed easily.

Right now, this dependence is disabled in both materials to
avoid side effects on the calculations that Jenn (jmfrede) is
performing.