LandIce: starting to put in logic to set 'Use Serial Mesh: false" automatically in the relevant input files in the case Trilinos was configured with
  -D TPL_Netcdf_PARALLEL:BOOL=OFF \

The idea is to increase the number of Albany tests that are run.
Because   -D TPL_Netcdf_PARALLEL:BOOL=OFF \ is off in Irina's nightly builds
(due to incompatibility of parallel netcdf with Fedora) most of the FELIX
tests are not run there.  They should be to help prevent breakage.