This commit should fix the Heat3DPUMI_Tpetra_RegressFail test.
The logic was not correct in Main_SolveT.cpp for the case when there
were no parameters.

We should write a non-SCOREC test that tests the WILL_FAIL logic
in the case that there are no parameters.

There is a possibility that new failures will arise in tomorrow's nightlies
for problems with no parameters, as it turns out comparisons were not being
performed in that case.  I will report tomorrow if that is the case.