Change in PHAL ETI macros policy
More specifically, the ETI macros that instantiate classes that depend
on 1+ scalar types should always allow RealType as scalar type, even if
the scalars of an evaluation type are all Fad/TanFad. This allows the user
to define pure input fields, that do NOT need any sort of derivative (wrt
solution or parameters).

Note: this does not apply to the ETI *_WITH_INPUT_OUTPUT_TYPES_*, since
that macro REQUIRES the output scalar type to be constructible from the
input one.

Note: this change led to compilation errors, since the interpolation of a
RealType field can still yield a FadType field, in case MeshScalarT is
indeed a FadType (as it happens for non-Residual evaluation types in the
case where the mesh depends on solution/parameters). Hence, for these
evaluators, some meta-programming logic is necessary to establish the output
scalar type.