LandIce: always request surf grad interpolation
This fixes some builds that enable CISM, but still want to test
standalone LandIce. So far, StokesFOBase would not build an
evaluator for the surface height gradient if CISM was enabled,
assuming that CISM would provide one. However, the field provided
by CISM and the one computed with the usual DOFGradInterpolation
evaluator have different names, so we can safely build both
evaluators, and let PHX discard the one that is not needed.