Per issue #214 discussion: switching nightlies on camobap to pull DTK version of the day.
We were using releave v2 in the nightlies apparently - I believe at some
point we switched to it b/c there were instabilities in the version of the day.

With a new DTK, we don't need the ```-D Tpetra_INST_INT_UNSIGNED_LONG:BOOL=ON``` to
compile with ETI on - it'll work out of the box!