Added a lightweight class that provides graph-like functionalities.
This class hides the concept of a graph, and only allows to create an
operator (and not to retrieve the underlying graph). If one really wants
the graph stored inside, then first create the operator, then cast to the
concrete linear algebra type, then extract the graph. In principle, you
should *never* need to do this.