Switching Epetra_GO to int by default.
Epetra64 seems to be not really supported in at least ML, Ifpack, AztecOO.
So we switch back to int for the epetra global ordinal. However, this means
that we can no longer take a GO* and reinterpret_cast-it to Epetra_GO*.
So all the Epetra utils need to create temporaries.