Fixing epetra utils
The assumption that vs of (multi)vectors/linear_ops in
epetra builds are always storing an RCP to the epetra map
was wrong. In particular, the domain vs of multivectors
is *always* a DefaultSpmdVectorSpace. Such vs are however
usually small and, most importantly, locally replicated,
so we can rebuild them manually (no risk of losing LID<->GID
distribution). Furthermore, trilinos uses the routine
create_VectorSpace from ThyraEpetraAdapters in order to
create a vs from an RCP<const Epetra_Map>, and attaches
the RCP to the epetra map with a different name than
the Albany createVectorSpace routine does.