LCM: adding logic to enable turning off Schwarz through the
-DENABLE_SCHWARZ

configure option to facilitate refactor of the code to remove petra and replace with thyra.

There is a lot of Tpetra in the Schwarz model evaluators and helper code.  My strategy in
refactoring is to get basic LCM compiling running; then proceed to more fancy features like
Schwarz (which require basic LCM to work anyway).