Bug fixes in GenericSTKMeshStruct. Also, removing createGatherVectorSpace util
The utility was not correct for the only use case we had. And fixing it to
work for that case would have made slightly counter-intuitive for the
general case. We were better off rolling its implementation where we used
it.

The GenericSTKMeshStruct had a bug, in that the serial vs used to load
fields from file should be created from a non-overlapping vs, otherwise
we are going to grab the same GID many times. We could, of course, remove
duplicates, but it is just easy to create a owned vs and grab gids from
there...