Another fix to GenericSTKMeshStruct
Turns out the comment in the previous commit was not correct. The nodes
extracted from the owned_part are still overlapping. So at this point the
only rational thing to do is to extract the nodes like we were doing before,
build the vs, extract the gids, gather all of them, sort them, *make them
unique*, then build the serial vs. The part within *'s was the culprit.