LCM: this push fixes a bunch of LCM tests that were failing.
It seems that nodalDofManager.numComponents() does not return the
right value in the case of a ScalarFieldType field.  This fix resolves
the problem - I would like to investigate why nodalDofManager.numComponents()
doesn't always return the right value, which may lead to a different fix.

There are currently 18 LCM tests that are failing:

The following tests FAILED:
         16 - utHeliumODEs (SEGFAULT)
         23 - Serial_Dynamic_Cubes_Newmark_Piro (Failed)
         24 - Parallel_Dynamic_Cubes_Newmark_Piro (Failed)
         39 - KfieldBC_L2Proj (Failed)
         40 - AnisotropicHyperelasticDamage_uniaxial (Failed)
         41 - AnisotropicHyperelasticDamage_shear (Failed)
         42 - AnisotropicHyperelasticDamage_hydrostatic (Failed)
         43 - AnisotropicDamage-Bifurcation_uniaxial (Failed)
         44 - AnisotropicDamage-Bifurcation_shear (Failed)
         45 - Gurson_uniaxial (Failed)
         46 - Gurson_shear (Failed)
         47 - Gurson_hydrostatic (Failed)
         48 - Neohookean_uniaxial (Failed)
         49 - Neohookean_shear (Failed)
         50 - Neohookean_hydrostatic (Failed)
         51 - Neohookean_general (Failed)
         52 - CrystalPlasticity_MPS_uniaxial (Failed)
         79 - RigidBody (Failed)

Tests 16 and 40-52 have a seg fault stemming from accessing a workset.stateArrayPtr
(e.g., line 128 of HeliumODEs_Def.hpp).

Test 23-24 are obsolete and should be removed from master as well as this branch.

Tests 39 and 79 utilize the ProjectIPtoNodalField response, which I am in the process
of debugging.

Note that when LCM is on, a bunch of basic tests fail, e.g., SteadyHeat1D, with bad
cast errors.  It's curious that this happens only when LCM is on, even though
the tests don't have to do with LCM.  I guess there is some logic that gets enabled
when LCM is on that all the tests use.  We need to debug this issue. To clarify,
it is not related to this commit (happens in the branch w/o the commit simply
if you turn on LCM).