LCM ACE: Convert thermal problem to dynamic.
Use the real (Tempus) time integrator to advance the thermal
problem in time. This requires converting the mechanics problem
to dynamic too. Adjust material models for this, and eliminate
the ad-hoc time integration for ACE heat conduction.

Add test in the multi-field STK field container to check that
when using multiple variables, all of them have the same number
of time derivatives. Otherwise the code fails with a segmentation
fault when trying to access inexistent derivatives for a field that
does not have them.