Bug fixes to CAS manager new aura functionalitites
When computing the aura gids specs (owners/users), we need to be careful.
In particular, if a rank owns 0 gids in one aura (ghosted or owned), the
check on the size of the gids specs would make the CAS manager try to
compute them (it would think it's the first call). However, the derived
classes *may* use global operations when computing the gids specs (it was
the case before this commit). If the call to such methods is triggered
a second time on a rank but not on all the other ranks, this can cause
MPI issues. To avoid this problem I
 - avoided globals in the Tpetra and Epetra derived classes
 - use stored flags to memorize the status of the specs, so that the methods
   to compute them are guaranteed to *not* be called multiple times.