Replacing 'Tpetra' with 'NONE' for the 'Build Type' default
Without this mod, when one runs an Epetra build and forgets to specify it in
the input file, we get errors quite a bit down the road (when the matrices
and vectors are fed to the linear solver, but they are Tpetra instead of
Epetra). The error message can be confusing, and hard to track down.

With this mod, one gets a clear error right away at the beginning of main,
where we set the build type, since 'NONE' is not a valid choice.