Mysterious bug fix in the Epetra CAS manager.
The RemoteIDList call on the owned aura epetra map was yielding some
nonsens PIDs. However, removing the pointer for the LIDs in the function
call (replacing it with nullptr) fixed this behavior. I am not sure why
this happens. I'm not satisfied with a fix that does not explain what
the error was. If you read this, keep an eye open in the CAS manager
for distributed memory errors.