LandIce testing: decomposing gis_unstruct_2d.exo on the fly if needed
That happens if Iopx is not found, AND mpi is enabled, AND tests use more
than 1 processor.

The tests dependencies have been fixed, so that not only mesh setup is
happening before the tests that need the mesh are run, but also if the
setup fails, those downstream tests are not run.