LCM: turning off DTK in debug build on camobap.
The failure in the SchwarzBar test that shows up in some of the nightlies
is in the non-DTK run of this problem; it doesn't show up with DTK.
The problem is in the getCoordinates() call here

SchwarzBC_Def.hpp:219

Here is the error: http://cdash.sandia.gov/CDash-2-3-0/testDetails.php?test=4594439&build=85216 .

I will look more at it tomorrow.  Any thoughts are appreciated.