Removed createLocallyReplicatedVectorSpace that takes array of gids
This routine was not conformant to Tpetra expectations (and caused exceptions
in a debug build). It was used only once, in the SolutionValuesResponseFunction.
Since the VS is not seen as 'replicated', we need to be careful with global
dimensions, otherwise some checks in Thyra/Tpetra may fail. We work around
this issue by using temporaries with the correct maps, and then locally
copy results into the 'really replicated' vectors.

Fixes #484