Bug fix in ResponseSquaredL2DifferenceSide
Also, changed (and upgraded) the way we establish if the target is zero.
Rather than check if 'Target Field Name' is 'ZERO' we now assume the
target is indeed a field if 'Target Field Name' is present (whatever it
is), otherwise look for 'Target Value'. Throw if neither or both are
present (mutually exclusive options).