CMake: limiting use of `include_directories', and adding info to installed cmake file
The installed cmake file now also provides a list of albany's tpls, as well
as information on their location.

Using include_directories at a top level directory is not advised, since
it automatically propagate to all the subdirectories added. On the other
hand, it is ok to use it at inner directories.
At the top level, prefer to use target-specific settings, unless you are
*absolutely certain* all targets will need that setting.