Hard-coding name of host skybridge-login5 in CDash submission.
Unfortunately this is the only way to get the code to push
correctly to the CDash site, which expects the nightlies to run on login5.
I am going to switch the nightlies to run on a compute node, which would
have an arbitrary name.  This is only relevant for pushing to Cdash.

Issue #485.