LCM ACE: Remove scaling from SDBC.
For mesh adaptation, the node sets need fixing for the SDBCs
due to BCs being applied to unknowns that may no longer
exist in the mesh.

In preparation to try to fix this, remove the scaling of
SDBCs, as it complicates things and it doesn't really
apply to SDBCs anyway.