Switching Cori nightly to run Cism-Albany tests, which I plan to move from camobap to cori.
Every time I do an OS upgrade on camobap, Cism-Albany gets messed up,
I believe due to netcdf incompatibilities that have nothing to do with the
code.  I don't have the time to deal with these every time I upgrade.
It makes more sense to run Cism-Albany on Cori anyway, as that is
a targeted platform for the code.