Adding missing ifdef guards.
The changes to PiroObserver aren't ready to go into master b/c they cause some
tests to fail, e.g.:

 SCOREC_MechWithTemp_Unif_Tpetra
 FO_Thermo
 Enthalpy

on some platforms.  To get the right time to the IPtoNodalField responses we need to call setCurrenTime in the PiroObserver,
which messes up something for the 3 problems above - need to look into.