LCM ACE: Make boundary parallel consistent
The boundary is now the same for the test case
in serial and parallel, and the erosion algorithm
gives the same result too. But the parallel case
crashes toward the end of the run when using
preconditioners (MueLu or Ifpack2), or gets
stuck in an apparent infinite loop if no
preconditioner is used.