Fixing the construction of Teuchos_Comm from MPI_Comm
The Teuchos::MpiComm class performs a collective operation
if the user does not specify a comm tag. This may cause the
program to hang if by any change the teuchos comm is not
created by all ranks in the MPI_Comm. To avoid this, we
construct the Teuchos::MpiComm with a default tag, large
enough to avoid potential clashes with actual message
tags that the user may use.

Also, fixed the return type of the routine that creates
the Teuchos comm, so that the RCP is templated on
const Teuchos_Comm.