Updating the nightlies that I own to set the following flags to use a static graph:
-D Tpetra_ENABLE_DEPRECATED_CODE:BOOL=OFF \
-D Xpetra_ENABLE_DEPRECATED_CODE:BOOL=OFF \

Mauro will check the dashboards for failures.

Note: I was unable to update the scripts on Cori, as I am unable to log on to the machine right now.
I've opened a ticket on this and will update them as soon as I can get on the machine.

Note2: the algol and fedora nightlies are owned by Alejandro, so those did not get updated.