Using a multiple-factories pattern to build the problem in Application
This way, we can hide details of downstream libraries in the main albany libs.

The downstream libraries details need to be pulled in before the problem
is created though. For this, we added a 'register_pb_factories' free function,
which looks at the enabled packages, and pulls in their factories.

Note: this free function is in a header that should *not* be shipped within
      the library. It is only used in the Main_*.cpp files. Downstream
      applications should perform similar operations to make sure their
      desired problem factories are enabled *before* they create an
      Albany::Application.