Introduced Albany::EpetraFECrs[Matrix|Graph] classes
These classes are *very* ligth weight. The only thing they do is to
store and expose the overlapped row map used at construction.

The reason we need these is that Epetra_FECrsMatrix does not
expose any information about the overlapped row map
(which is only computed during GlobalAssemble).
In Albany, we need that info, since our thyra utility to modify
matrix entries only works with local ids (for both row and
column indices). We could make a utility that works with global
ids, but then we would need to convert every time to local ids
when using Tpetra. In this respect, Tpetra and Epetra are very
different.

Since we are forced to have one impl (T or E) converting the
ids, we decide to have Epetra do it (since it's less used).
However, to do that we need to store the map of the overlapped
distribution. Therefore, we create lightweight classes,
derived from the Epetra FE ones, which store and expose this
extra bit of information.

Note: if only Epetra_FECrsGraph exposed the overlapped map, AND
      Epetra_FECrsMatrix exposed the nonlocal graph, we would
      not need this.