Bug fix in thermal problem with source term.
Creating a forced thermal problem to be used to verify transient sensitivites on.
The problem has an exact closed-form analytical solution. Convergence
of the Albany solution with respect to the exact solution has been verified
in space as well as time (w/o sensitivities for now).

Switching thermal problems to parallel and to use MueLu instead of Ifpack2
for preconditioner.  It's kind of curious that MueLu is needed for linear solve
to not fail for this problem when run in parallel.