Creating ComputeNumberParams() routine in Albany::Utils.
This allows one to remove passing of (redundant with the PL) numParams
input for the AAdapt::AdaptiveSolutionManager and Albany::DiscretizationFactory
constructors.

Note that numParams still needs to be passed to the STKMeshStruct constructors,
but there is no redundancy there with the PL passed in (the PL is discParams only
which has no info about the number of params).