LandIce: refactoring of NullSpaceUtility
For scalar equations it often makes sense to have constants in the near null space
For elesticity type problems, the null space is constituted of rigid body modes,
 consisting of constants/transalations and rotations associated to a physical vector
Here we assume that the we have a set of equations and possibly the the first 2 or 3  equations
 are for a 2d or 3d physical vector (e.g. velocity).
We can compute constant modes for all the equations and then add rotations for the vector.
It is possible to choose whether to compute constant modes and/or rotation modes.
The dimension of the null space is determined by the class and does not need to be computed by the user.
Different checks have beed added to make sure that the methods of the RigidBodyModes class
 are called in the right sequence.
Also we check whether the interleaved ordering is compatible with what implemented.