LandIce: Remove the flag ENABLE_USE_CISM_FLOW_PARAMETERS
The option refers to the coefficients in the formula of  flow factor
as a function o the temperature.
The option was used to allow to choose between CISM coefficiens and
those used in LifeV. In recent years we only used CISM coefficients
(which are the same adoped by MPAS) so there is no need to provide
that option any longer.