Uniforming Albany's kokkos types to those used by Tpetra
A recent change in Trilinos (pr 7677) changed the template arguments
of kokkos local structures (graph and matrix) used by Tpetra's Crs classes.

This commit makes our template types equal to those used in Tpetra,
and also adds a static assert verifying that, to make it easy to track
the problem if things change again in Trilinos in the future.