Disabling a debug check in Tpetra
The check is in FECrsGraph, and it checks that, if you own an element,
then you must own at least one of its nodes. Such check ensures that the
column map of the (owned) matrix is as small as possible, but it is too
restrictive for our stk (and MPAS) meshes. The best thing to do would be
to fix the node sharing (or, better, the element sharing) properties in
the stk bulk data structures.