Refactored SaveSideSetState and LoadSideSetState
The two ev now rely directly on STK structure, since it makes
it easier to loop over entities, and allows to avoid using
stuff like state arrays (which are Shards arrays).

Also, for SaveSideSetState, we use FieldRankType and FieldLocation
(passed in through the input parameter list) to deduce the field
layout. For rank 'Gradient' (supported only if location is Cell),
we do not save the field V, but rather the product J*V, where J
is the matrix of the tangent vectors (a 3x2 matrix for a 3d volume mesh).

Big caveat: since the field layout has dimension 2 along the Dim tag
(for a 3d volume mesh), we cannot store J*V itself, since J*V has
length 3. However, since virtually all our use cases feature a side
mesh that is in the xy plane, we simply "ignore" the 3rd component
of J*V. This is a hack, that works for now. We will come up with
a more elegant way later (e.g., when registering the state, if the
field rank is Gradient, we have a phx field of length 2, but the
stk field has length 3, that is, the dim of the ambient space).