Switch to current correct TriBITS usage of libraries for TeuchosParameterList
This Albany CMakeLists.txt was making assumptions about the implentation
details of the TriBITS-generated <Package>Config.cmake files that it should
not have been making.  It was assuming that the raw TriBITS target
'teuchosparameterlist' existed which is a no-no.  The correct old-school
TriBITS usage is ${TeuchosParameterList_LIBRARIES} and the associated include
directories.

This works with old TriBITS and refactored TriBITS.

This is related to trilinos/Trilinos#9972 and PR trilinos/Trilinos#9978.