Refactor of the Hessian response for allowing for custom designed solvers
Until now the Hessian pp matrix has always been a Crs Matrix that can be solved for using standard solvers.
However, in some cases it is not easy or performant to explicitly build the matrix.
This commit allows a custom linear operator and solves for each response.
If such operator is not provided, the Hessian response matrix is built using the current coloring-seeding approach that relies on Hessian vectors products
and on the assumption of a finite-element like pattern (for distributed parameters)

This commit also implements the solver for the quadratic response.

A few tests are modified to comply with the changes.