enable mpi, using omegah with pure kokkos backend
MPIEXEC_EXECUTABLE gives hints to cmake find_package(MPI) on where to find the MPI install.  In most cases I expect the MPI C++ compiler wrapper to be passed from Albany as the C++ compiler, but setting this cmake arg adds some insurance when/if it is not being used.