Allow to not save solution field in output mesh
This is so far only useful for the PopulateMesh problem, where
setting the solution field in the mesh would cause problems when
loading the mesh later on in another problem. That is because
STK forces redeclarations of the same field to have a compatible layout,
so we would only be able to use the populated mesh for a problem
that has the same number of eqns as we declared in the PopulateMesh
problem.